+++ src/Propulsion/Submission.fs
@@ -2,10 +2,10 @@
 
 open Serilog
 open System
-open System.Collections.Concurrent
 open System.Collections.Generic
 open System.Diagnostics
 open System.Threading
+open System.Threading.Tasks
 
 module Internal =
 
@@ -33,12 +33,45 @@ module Internal =
             let due = timer.ElapsedMilliseconds > max
             if due then timer.Restart()
             due
+    let timeRemaining (period : TimeSpan) =
+        let timer, max = Stopwatch.StartNew(), int64 period.TotalMilliseconds
+        fun () ->
+            match max - timer.ElapsedMilliseconds |> int with
+            | rem when rem <= 0 -> timer.Restart(); true, max
+            | rem -> false, rem
+    let atTimedIntervals (period : TimeSpan) =
+        let timer, max = Stopwatch.StartNew(), int64 period.TotalMilliseconds
+        let remNow () = max - timer.ElapsedMilliseconds |> int
+        fun f ->
+            match remNow () with
+            | rem when rem <= 0 -> f (); timer.Restart(); remNow ()
+            | rem -> rem
+
+    module Channel =
+
+        open System.Threading.Channels
+
+        let unboundedSr<'t> = Channel.CreateUnbounded<'t>(UnboundedChannelOptions(SingleReader = true))
+        let unboundedSwSr<'t> = Channel.CreateUnbounded<'t>(UnboundedChannelOptions(SingleWriter = true, SingleReader = true))
+        let write (c : Channel<_>) = c.Writer.TryWrite >> ignore
+        let awaitRead (c : Channel<_>) ct = let vt = c.Reader.WaitToReadAsync(ct) in vt.AsTask()
+        let apply (c : Channel<_>) f =
+            let mutable worked, msg = false, Unchecked.defaultof<_>
+            while c.Reader.TryRead(&msg) do
+                worked <- true
+                f msg
+            worked
+
+    module Task =
+
+        let start create = Task.Run<unit>(Func<Task<unit>> create) |> ignore<Task>
 
     type Sem(max) =
         let inner = new SemaphoreSlim(max)
         member _.HasCapacity = inner.CurrentCount <> 0
         member _.State = max-inner.CurrentCount,max
         member _.Await(ct : CancellationToken) = inner.WaitAsync(ct) |> Async.AwaitTaskCorrect
+        member x.AwaitButRelease() = inner.WaitAsync().ContinueWith(fun _t -> x.Release())
         member _.Release() = inner.Release() |> ignore
         member _.TryTake() = inner.Wait 0
 
@@ -60,47 +93,46 @@ module Submission =
     /// Holds the stream of incoming batches, grouping by partition
     /// Manages the submission of batches into the Scheduler in a fair manner
     type SubmissionEngine<'S, 'M, 'B when 'S : equality>
-        (   log : ILogger, maxSubmitsPerPartition, mapBatch : (unit -> unit) -> SubmissionBatch<'S, 'M> -> 'B, submitBatch : 'B -> int, statsInterval, ?pumpInterval : TimeSpan,
+        (   log : ILogger, maxSubmitsPerPartition, mapBatch : (unit -> unit) -> SubmissionBatch<'S, 'M> -> 'B, submitBatch : 'B -> int, statsInterval,
             ?tryCompactQueue) =
 
-        let pumpInterval = defaultArg pumpInterval (TimeSpan.FromMilliseconds 5.)
-        let incoming = new BlockingCollection<SubmissionBatch<'S, 'M>[]>(ConcurrentQueue())
+        let awaitIncoming, applyIncoming, enqueueIncoming =
+            let c = Channel.unboundedSr
+            Channel.awaitRead c, Channel.apply c, Channel.write c
         let buffer = Dictionary<'S, PartitionQueue<'B>>()
+
         let mutable cycles, ingested, completed, compacted = 0, 0, 0, 0
         let submittedBatches,submittedMessages = PartitionStats(), PartitionStats()
-
+        let statsInterval = timeRemaining statsInterval
         let dumpStats () =
             let waiting = seq { for x in buffer do if x.Value.queue.Count <> 0 then yield struct (x.Key, x.Value.queue.Count) } |> sortByVsndDescending
-            log.Information("Submitter {cycles} cycles {ingested} accepted {compactions} compactions Holding {@waiting}", cycles, ingested, compacted, waiting)
-            log.Information(" Submitted {@batches} Completed {completed} Messages {@messages}", submittedBatches.StatsDescending, completed, submittedMessages.StatsDescending)
+            log.Information("Submitter ingested {ingested} compacted {compacted} completed {completed} Events {items} Batches {batches} Holding {holding} Cycles {cycles}",
+                            ingested, compacted, completed, submittedMessages.StatsDescending, submittedBatches.StatsDescending, waiting, cycles)
             cycles <- 0; ingested <- 0; compacted <- 0; completed <- 0; submittedBatches.Clear(); submittedMessages.Clear()
-
-        let maybeLogStats =
-            let due = intervalCheck statsInterval
-            fun () ->
-                cycles <- cycles + 1
-                if due () then dumpStats ()
+        let maybeDumpStats () =
+            cycles <- cycles + 1
+            let due, remaining = statsInterval ()
+            if due then dumpStats ()
+            int remaining
 
         // Loop, submitting 0 or 1 item per partition per iteration to ensure
         // - each partition has a controlled maximum number of entrants in the scheduler queue
         // - a fair ordering of batch submissions
-        let propagate () =
-            let mutable more, worked = true, false
-            while more do
-                more <- false
-                for KeyValue (pi, pq) in buffer do
-                    if pq.queue.Count <> 0 then
-                        if pq.submissions.TryTake() then
-                            worked <- true
-                            more <- true
-                            let count = submitBatch <| pq.queue.Dequeue()
-                            submittedBatches.Record(pi)
-                            submittedMessages.Record(pi, int64 count)
+        let tryPropagate (waiting : ResizeArray<Sem>) =
+            waiting.Clear()
+            let mutable worked = false
+            for KeyValue (pi, pq) in buffer do
+                if pq.queue.Count <> 0 then
+                    if pq.submissions.TryTake() then
+                        worked <- true
+                        let count = submitBatch <| pq.queue.Dequeue()
+                        submittedBatches.Record(pi)
+                        submittedMessages.Record(pi, int64 count)
+                    else waiting.Add(pq.submissions)
             worked
 
-        /// Take one timeslice worth of ingestion and add to relevant partition queues
-        /// When ingested, we allow one propagation submission per partition
         let ingest (partitionBatches : SubmissionBatch<'S, 'M>[]) =
+            ingested <- ingested + 1
             for { source = pid } as batch in partitionBatches do
                 let pq =
                     match buffer.TryGetValue pid with
@@ -111,7 +143,6 @@ module Submission =
                     pq.submissions.Release()
                 let mapped = mapBatch markCompleted batch
                 pq.Append(mapped)
-            propagate()
 
         /// We use timeslices where we're we've fully provisioned the scheduler to index any waiting Batches
         let compact f =
@@ -119,31 +150,26 @@ module Submission =
             for KeyValue(_, pq) in buffer do
                 if f pq.queue then
                     worked <- true
-            if worked then compacted <- compacted + 1; true
-            else false
+            if worked then compacted <- compacted + 1
+            worked
+        let maybeCompact () =
+            match tryCompactQueue with
+            | Some f -> compact f
+            | None -> false
 
         /// Processing loop, continuously splitting `Submit`ted items into per-partition queues and ensuring enough items are provided to the Scheduler
-        member _.Pump() = async {
-            let! ct = Async.CancellationToken
+        member _.Pump(ct : CancellationToken) = task {
+            // Semaphores for partitions that have reached their submit limit; if capacity becomes available, we want to wake to submit
+            let waitingSubmissions = ResizeArray<Sem>()
+            let submitCapacityAvailable : seq<Task> = seq { for w in waitingSubmissions -> w.AwaitButRelease() }
             while not ct.IsCancellationRequested do
-                let mutable items = Unchecked.defaultof<_>
-                let mutable propagated = false
-                if incoming.TryTake(&items, pumpInterval) then
-                    propagated <- ingest items
-                    while incoming.TryTake(&items) do
-                        if ingest items then propagated <- true
-                else propagated <- propagate ()
-                match propagated, tryCompactQueue with
-                | false, None -> Thread.Sleep 2
-                | false, Some f when not (compact f) -> Thread.Sleep 2
-                | _ -> ()
-
-                maybeLogStats () }
+                while applyIncoming ingest || tryPropagate waitingSubmissions || maybeCompact () do ()
+                let nextStatsIntervalMs = maybeDumpStats ()
+                do! Task.WhenAny[| awaitIncoming ct :> Task; yield! submitCapacityAvailable; Task.Delay(nextStatsIntervalMs) |] :> Task }
 
         /// Supplies a set of Batches for holding and forwarding to scheduler at the right time
         member _.Ingest(items : SubmissionBatch<'S, 'M>[]) =
-            Interlocked.Increment(&ingested) |> ignore
-            incoming.Add items
+            enqueueIncoming items
 
         /// Supplies an incoming Batch for holding and forwarding to scheduler at the right time
         member x.Ingest(batch : SubmissionBatch<'S, 'M>) =